user-service/requirements.txt
Flask
Flask-Bcrypt
pymongo
PyJWT
gunicorn
python-dotenv
